fix(publish): review corrections — dispatch env gate, docstring hygiene, helper dedup
(1) the publish job environment resolves to none on workflow_dispatch so the
build-only dry-run isn't blocked by the tag-restricted deployment policy;
(2) remove milestone/stage attribution from 4 test docstrings (§7a);
(3) drop the current_image_id(service_name) generalization + shim, keep the
single current_runner_image_id; (4) dedup the health-poll helper into
docker_helpers.
