+++ tests/utils/docker_helpers.py
@@ -11,6 +11,9 @@ This module provides two distinct Docker availability checks:
   manage containers directly via the Docker SDK.
 """
 
+import subprocess
+import time
+
 import pytest
 
 # ---------------------------------------------------------------------------
@@ -19,6 +22,9 @@ import pytest
 
 DOCKER_RUNNER_ADDRESS = "localhost:50051"
 DOCKER_DB_SERVICE_URL = "http://localhost:8000"
+
+HEALTHY_TIMEOUT_S = 120.0
+_HEALTH_POLL_INTERVAL_S = 2.0
 
 
 # ---------------------------------------------------------------------------
@@ -75,11 +81,11 @@ def is_docker_daemon_available() -> bool:
         return False
 
 
-def current_image_id(service_name: str) -> str | None:
-    """Docker id of *service_name*'s image the *current tree* produces, or ``None``.
+def current_runner_image_id() -> str | None:
+    """Docker id of the runner image the *current tree* produces, or ``None``.
 
-    Resolves the exact content-hash ref ``builder.expected_image_ref(service_name)``
-    — the tag a real ``build_image(service_name)`` assigns — and inspects that one
+    Resolves the exact content-hash ref ``builder.expected_image_ref("runner")``
+    — the tag a real ``build_image("runner")`` assigns — and inspects that one
     ref. It is a single exact-ref lookup: no ``docker images`` enumeration, no
     ``.Created`` ranking, no candidate filtering. A foreign image built from a
     different Dockerfile hashes to a different tag, so it cannot match the ref;
@@ -89,11 +95,9 @@ def current_image_id(service_name: str) -> str | None:
     case). The full ``sha256:...`` digest is returned so it compares equal to
     the Docker SDK's ``Image.image_id``.
     """
-    import subprocess
-
     from tolokaforge.docker import builder
 
-    ref = builder.expected_image_ref(service_name)
+    ref = builder.expected_image_ref("runner")
     result = subprocess.run(
         ["docker", "image", "inspect", ref, "--format", "{{.Id}}"],
         capture_output=True,
@@ -104,6 +108,23 @@ def current_image_id(service_name: str) -> str | None:
     return result.stdout.strip() or None
 
 
-def current_runner_image_id() -> str | None:
-    """Docker id of the runner image the *current tree* produces, or ``None``."""
-    return current_image_id("runner")
+def wait_for_health(container_id: str, timeout_s: float = HEALTHY_TIMEOUT_S) -> str:
+    """Poll ``container_id`` Docker health until ``healthy``/``unhealthy`` or timeout.
+
+    Returns the last observed status. ``unhealthy`` is returned as soon as it is
+    seen so a failing probe fails fast; an empty string means the container is
+    gone or reports no health (inspect failed).
+    """
+    deadline = time.monotonic() + timeout_s
+    status = ""
+    while time.monotonic() < deadline:
+        result = subprocess.run(
+            ["docker", "inspect", "--format", "{{.State.Health.Status}}", container_id],
+            capture_output=True,
+            text=True,
+        )
+        status = result.stdout.strip()
+        if status in {"healthy", "unhealthy"}:
+            return status
+        time.sleep(_HEALTH_POLL_INTERVAL_S)
+    return status
